Added utils to landsat_processing_utils, started AFD algo

+++ landsat_processing_utils.py
@@ -53,6 +53,191 @@ def histogram_stretch(img, min_vals = None, max_vals = 99):
     
     return img_stretched
 
+def landsat_l2_scaling(stacked_mb_raster):    
+    '''
+    stacked_mb_raster: filepath to stacked l2 raster to be scaled (must be ordered bands 1-X then 10)
+        OR 3D stacked numpy array (must be ordered bands 1-X then 10)
+    
+    returns 3D numpy array of the scaled, stacked bands as well as the profile information
+    '''
+    if isinstance(stacked_mb_raster, str):
+        with rio.open(stacked_mb_raster) as src:
+            profile = src.profile
+            stack = src.read()
+            src.close()
+    else:
+        stack = stacked_mb_raster
+
+    # change the stack from DNs to reflectances
+    # remember temperature uses a different scaling factor (temp is in Kelvin)
+    scaled_stack = stack[:-1,...] * 0.0000275 - 0.2
+    scaled_temp = stack[-1,...] * 0.00341802 + 149.0 
+
+    # need to change scaled_temp from 2D to 3D
+    scaled_stack = np.vstack((scaled_stack, np.expand_dims(scaled_temp, axis = 0))) 
+    
+    return scaled_stack, profile
+
+def prep_training_data(raster, gdf, scale = False):
+    '''
+    Prepares X_train and y_train data from a raster image given a GeoDataFrame of classified polygons
+    
+    raster: Path to stacked 3D raster. If conducting scaling, 3D raster should be in the form (Bands 1-X, 10).
+    
+    gdf: GeoDataFrame containing labelled polygons. Must have "geometry" col and the labels col must be named "labels"
+        May want to adjust this in the future.
+        
+    scale: Flag to indicate if the raster should be scaled from DNs to reflectances. Current implementation only works
+        with Landsat Level 2 scaling. Please run process_folder() to do TOA scaling on Landsat level 1 products prior to using
+        this function.
+    '''
+
+    # note that although we have scaled_stack, rio.mask.mask requires a dataset in read mode
+    with rio.open(raster) as src:
+        profile = src.profile
+        
+        # To prep the data for ML analysis, we need two numpy arrays:
+        # X: a numpy array that contains all of the band data for the pixel
+        # y: the labels for training
+
+        # Creates an empty array with X columns, where X is the number of bands in the multiband raster
+        X_train = np.array([], dtype = np.float32).reshape(0, profile["count"])
+        y_train = np.array([], dtype = np.string_) # labels for training
+        
+        # Iterate over each polygon in our landuse ground truth dataset
+        for index, row in gdf.iterrows():
+            feature = [row["geometry"]]
+
+            # crop the image - mask function returns a tuple
+            out_image, out_transform = rio.mask.mask(src, feature, crop = True)
+
+            # out_image has a shape (8, height, width)
+            # Since this returns a rectangular array, and our shape is not rectangular, there will be
+            # a bunch of nodata - get rid of them.
+
+            # ~np.any(np.isnan(out_image), axis = 0) if any of the bands are np.nan
+
+            # note that this gets rid of any column that has ANY nans
+            # The following code returns a column for each band, with each row representing a pixel
+            out_image_trimmed = out_image[:, ~np.any(out_image == profile['nodata'], axis = 0)]
+
+            # We actually want this the other way around - we want a row for pixel, and a column for each
+            # band - so we transpose the image
+            out_image_trimmed = out_image_trimmed.T
+            
+            if scale:
+                out_bands = out_image_trimmed[:,:-1] * 0.0000275 - 0.2
+                out_temp = out_image_trimmed[:,-1] * 0.00341802 + 149.0 
+
+                out_raster = np.hstack((out_bands, np.expand_dims(out_temp, axis = 1)))
+            else:
+                out_raster = out_image_trimmed
+            
+            # We append the labels to the answer array equal to the number of pixels:
+            # Remember to put brackets around the row["LULC"], or else you'll get "forestforestforest"
+            # out_image_trimmed.shape[0] is the number of pixels in the training data
+            y_train = np.append(y_train, [row["labels"]] * out_raster.shape[0])
+
+            # vstack is like concat for rows. Note that to vstack correctly, the array tuple that you feed
+            # to vstack must have the same column dimension
+            X_train = np.vstack((X_train, out_raster))
+
+    src.close()
+    
+    return X_train, y_train
+
+# Run a conservative cloud detection
+def qa_pixel_interp_aggressive(number):
+    '''
+    Helps interpret the 16bit data in the landsat qa pixels
+    
+    returns True if there is mid confidence cirrus, snow/ice, cloud shadow, OR clouds
+    '''
+    binary = bin(number)[2:].zfill(16)
+    
+    # if medium to high confidence cirrus, snow/ice, cloud shadow, and clouds
+    if int(binary[:2]) > 1:
+        return True
+    elif int(binary[2:4]) > 1:
+        return True
+    elif int(binary[4:6]) > 1:
+        return True
+    elif int(binary[6:8]) > 1:
+        return True
+    else:
+        return False
+    
+def qa_pixel_interp_conservative(number):
+    '''
+    Helps interpret the 16bit data in the landsat qa pixels
+    
+    returns True if there is mid confidence cirrus, snow/ice, cloud shadow, OR clouds
+    '''
+    binary = bin(number)[2:].zfill(16)
+    
+    # if high confidence cirrus, snow/ice, cloud shadow, and clouds
+    # 01 - low, 10 - medium, 11 - high
+    if int(binary[:2]) > 10:
+        return True
+    elif int(binary[2:4]) > 10:
+        return True
+    elif int(binary[4:6]) > 10:
+        return True
+    elif int(binary[6:8]) > 10:
+        return True
+    else:
+        return False
+    
+def qa_pixel_interp_conserv_water(number):
+    '''
+    Helps interpret the 16bit data in the landsat qa pixels
+    
+    returns True if there is mid confidence cirrus, snow/ice, cloud shadow, clouds OR WATER
+    '''
+    binary = bin(number)[2:].zfill(16)
+    
+    # if high confidence cirrus, snow/ice, cloud shadow, and clouds
+    # 01 - low, 10 - medium, 11 - high
+    if int(binary[:2]) > 10:
+        return True
+    elif int(binary[2:4]) > 10:
+        return True
+    elif int(binary[4:6]) > 10:
+        return True
+    elif int(binary[6:8]) > 10:
+        return True
+    # if water return true
+    elif int(binary[8]) == 1:
+        return True
+    else:
+        return False
+    
+def apply_array_func(func, x):
+    '''
+    Applies a function element-wise across a 1D array
+    '''
+    return np.array([func(xi) for xi in x])
+
+def run_qa_parser(qa_raster, func):
+    '''
+    Accepts a QA_PIXEL array bundled with Landsat l2 product consisting of 16 bit unsigned integers.
+    Generates a binary cloud mask using the function provided.
+    
+    qa_raster: numpy array of qa_raster. Works both stacked and unstacked.
+    func: interpretation scheme to be used to generate the cloud mask. Current schemes include:
+        qa_pixel_interp_conserv_water: conservative cloud detection, also masks water bodies
+        qa_pixel_interp_conservative: conservative cloud detection
+        qa_pixel_interp_aggressive: aggressive cloud detection
+    
+    Returns a squeezed binary cloud mask
+    '''
+    unique_vals = np.unique(qa_raster)
+    masked_vals = apply_array_func(func, unique_vals)
+    masked_vals = unique_vals[masked_vals]
+    cl_mask = np.isin(qa_raster, masked_vals)
+    
+    return cl_mask.squeeze()
+
 
 # In[39]:
 
@@ -186,7 +371,8 @@ def earthpy_cropper(filenames, target_shapefile, profile, cleanup = False):
     # only uncomment if you are SURE YOU WANT TO DO THIS
     if cleanup:
         for file in filenames:
-            os.remove(file)
+            if "TOA" in file:
+                os.remove(file)
             
     return band_paths
 
@@ -307,7 +493,7 @@ if __name__ == "__main__":
     parser.add_argument('-s', '--stack', action='store_true', help = 'Create a stacked raster of the images - default False') 
     parser.add_argument('-c', '--sun_corr', action='store_false', help = "add this flag if you DON'T want to do a sun elevation correction - default True")
     parser.add_argument('-o', '--outdir', help = "Specify an output folder to save TOA corrected images")
-    parser.add_argument('-d', '--cleanup', action='store_true', help = "If cropping, choose whether to delete the uncropped TOA image - default False")
+    parser.add_argument('-d', '--cleanup', action='store_true', help = "If cropping, choose whether to delete the uncropped TOA images - default False")
 
 
     ### Preview arguments
@@ -328,7 +514,7 @@ if __name__ == "__main__":
 # In[44]:
 
 
-parser.parse_args('-m "study_area.geojson" -s -c -o "20230710_TOA" "LC08_L1TP_180035_20230710_20230718_02_T1"'.split(' '))
+# parser.parse_args('-m "study_area.geojson" -s -c -o "20230710_TOA" "LC08_L1TP_180035_20230710_20230718_02_T1"'.split(' '))
 
 
 # In[46]:
